+++ examples/lazy-load/src/index.js
@@ -1,0 +1,10 @@
+import {startComponents} from '@gondel/core';
+
+import './components/button';
+
+startComponents();
+
+
+
+
+