+++ examples/typescript/src/index.ts
@@ -1,12 +1,9 @@
-import { startComponents } from "@gondel/core";
-import { initEventPlugin } from "@gondel/plugin-events";
+// Optional hot module reloading:
 import { hot } from "@gondel/plugin-hot";
 declare const module: any;
 hot(module);
-initEventPlugin();
 
-// Load all components
-const componentsContext = (require as any).context("./components", true, /\.ts$/);
-componentsContext.keys().forEach(componentsContext);
-
-startComponents();
+// Load components
+import "./components/button";
+import "./components/form";
+import "./components/input";