+++ examples/typescript/webpack.config.js
@@ -22,7 +22,7 @@ module.exports = {
   },
   plugins: [
     new HtmlWebpackPlugin({
-      template: 'src/index.html'
+      template: 'public/index.html'
     })
   ],
 };