+++ packages/core/src/GondelEventRegistry.ts
@@ -25,25 +25,39 @@ export type INamespacedEventHandlerRegistry = {
   };
 };
 
+/**
+ * The current context information neccessary to
+ * execute an event
+ */
+export type IEventExecutionContext = {
+  ctx: HTMLElement;
+  /**
+   * The current listener target
+   */
+  target: HTMLElement;
+  handlerOptions: Array<IHandlerOption>;
+};
+
 export type IHandlerOption = {
   selector?: string;
   handlerName: string;
 };
 
-const domEventRegistry: {
-  [eventName: string]: IEventHandlerRegistry;
-} =
-  (window as any).__gondelDomEvents || {};
-(window as any).__gondelDomEvents = domEventRegistry;
-
-/* istanbul ignore next : Browser polyfill can't be tested */
-const matches =
-  Element.prototype.matches ||
-  (Element as any).prototype.matchesSelector ||
-  (Element as any).prototype.mozMatchesSelector ||
-  (Element as any).prototype.msMatchesSelector ||
-  (Element as any).prototype.oMatchesSelector ||
-  (Element as any).prototype.webkitMatchesSelector;
+// Polyfill for element.prototype.matches
+let matchesCssSelector = (element: Element, selector: string): true => {
+  const elementPrototype = (window as any).Element.prototype;
+  /* istanbul ignore next : Browser polyfill can't be tested */
+  const elementMatches =
+    elementPrototype.matches ||
+    elementPrototype.matchesSelector ||
+    elementPrototype.mozMatchesSelector ||
+    elementPrototype.msMatchesSelector ||
+    elementPrototype.webkitMatchesSelector;
+  // Cache the function and call it
+  return (matchesCssSelector = (element: Element, selector: string): true => {
+    return elementMatches.call(element, selector);
+  })(element, selector);
+};
 
 function getParentElements(startElement: HTMLElement): Array<HTMLElement> {
   const nodes = [];
@@ -60,7 +74,7 @@ export function getHandlers(
   attributeName: string,
   eventHandlerRegistry: INamespacedEventHandlerRegistry,
   target: HTMLElement
-): Array<{ ctx: HTMLElement; handlerOptions: Array<IHandlerOption> }> {
+): Array<IEventExecutionContext> {
   const parents = getParentElements(target);
   // Find all selectors which have been registred for this event type
   // and which have a gondel component in one of the parrent nodes
@@ -82,31 +96,36 @@ export function getHandlers(
   const handlerQueue: Array<{
     index: number;
     ctx: HTMLElement;
+    target: HTMLElement;
     handlerOptions: Array<IHandlerOption>;
   }> = [];
   selectorsOfFoundComponents.forEach(({ index, handlers }) => {
     const selectorNames = Object.keys(handlers);
-    selectorNames.forEach((selectorName): any => {
-      // If no selector is given the handler does always match
-      if (!selectorName) {
-        return handlerQueue.push({
-          index,
-          ctx: parents[index],
-          handlerOptions: handlers[selectorName]
-        });
-      }
-      // Iterate backwards over the children of the component to find an element
-      // which matches the selector for the current handler
-      for (let i = index; --i >= 0; ) {
-        if (matches.call(parents[i], selectorName)) {
+    selectorNames.forEach(
+      (selectorName): any => {
+        // If no selector is given the handler does always match
+        if (!selectorName) {
           return handlerQueue.push({
-            index: i,
+            index,
             ctx: parents[index],
+            target: parents[index],
             handlerOptions: handlers[selectorName]
           });
         }
+        // Iterate backwards over the children of the component to find an element
+        // which matches the selector for the current handler
+        for (let i = index; --i >= 0; ) {
+          if (matchesCssSelector(parents[i], selectorName)) {
+            return handlerQueue.push({
+              index: i,
+              ctx: parents[index],
+              target: parents[i],
+              handlerOptions: handlers[selectorName]
+            });
+          }
+        }
       }
-    });
+    );
   });
   // Break if we couldn't find any matching element
   if (handlerQueue.length === 0) {
@@ -134,15 +153,23 @@ function handleEvent(
   executeHandlers(handlers, event, namespace);
 }
 
+let _domEventRegistry: {
+  [eventName: string]: IEventHandlerRegistry;
+};
+
 /**
  * Returns the namespace registry for the given namespace..
  * This function must be used only by core or plugins
  */
-export function getEventRegistry(namespace: string) {
-  if (!domEventRegistry[namespace]) {
-    domEventRegistry[namespace] = {};
+export function getEventRegistry(namespace: string): IEventHandlerRegistry {
+  if (!_domEventRegistry) {
+    _domEventRegistry = (window as any)["__\ud83d\udea1DomEvents"] || {};
+    (window as any)["__\ud83d\udea1DomEvents"] = _domEventRegistry;
   }
-  return domEventRegistry[namespace];
+  if (!_domEventRegistry[namespace]) {
+    _domEventRegistry[namespace] = {};
+  }
+  return _domEventRegistry[namespace];
 }
 
 /**
@@ -150,10 +177,13 @@ export function getEventRegistry(namespace: string) {
  * This function must be used by core or only by plugins
  */
 export function executeHandlers(
-  handlers: Array<{ ctx: HTMLElement; handlerOptions: Array<IHandlerOption> }>,
+  handlers: Array<IEventExecutionContext>,
   event: Event,
   namespace: string
 ) {
+  /** Store wether the original Event was modified to provide the correct currentTarget */
+  let eventObjectRequiresCleanup = false;
+  /** Store optional callback results which are executed together to allow grouped redraws */
   const results = [];
   for (let i = 0; i < handlers.length && !event.cancelBubble; i++) {
     const handlerObject = handlers[i];
@@ -161,6 +191,12 @@ export function executeHandlers(
     const gondelComponent = getComponentByDomNode(handlerObject.ctx, namespace);
     // Skip if the component wasn't started or if it was stopped
     if (gondelComponent) {
+      // See https://stackoverflow.com/questions/52057726/what-is-the-best-way-to-alter-a-native-browser-event
+      Object.defineProperty(event, "currentTarget", {
+        value: handlerObject.target,
+        configurable: true
+      });
+      eventObjectRequiresCleanup = true;
       for (let j = 0; j < handlerOptions.length && !event.cancelBubble; j++) {
         const handlerResult = (gondelComponent as any)[handlerOptions[j].handlerName].call(
           gondelComponent,
@@ -176,6 +212,11 @@ export function executeHandlers(
   results.forEach(result => {
     result();
   });
+  // Cleanup the event object
+  if (eventObjectRequiresCleanup) {
+    // See https://stackoverflow.com/questions/52057726/what-is-the-best-way-to-alter-a-native-browser-event
+    delete (event as any).currentTarget;
+  }
 }
 
 /**
@@ -189,7 +230,7 @@ function startListeningForEvent(eventName: string, namespace: string) {
       null,
       namespace,
       `data-${namespace}-name`,
-      domEventRegistry[namespace][eventName]
+      getEventRegistry(namespace)[eventName]
     )
   );
 }